Optimize FileFormat in hive sink
