+++ inlong-manager/manager-common/src/main/java/org/apache/inlong/manager/common/enums/FileFormat.java
@@ -1,0 +1,33 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.inlong.manager.common.enums;
+
+public enum FileFormat {
+
+    TextFile, ORCFile, SequenceFile, Parquet;
+
+    public static FileFormat forName(String name) {
+        for (FileFormat value : values()) {
+            if (value.name().equals(name)) {
+                return value;
+            }
+        }
+        throw new IllegalArgumentException(String.format("Unsupport FileFormat:%s", name));
+    }
+
+}
